fix: https://github.com/TypeFox/monaco-components/issues/64 service initalization can only be triggered once
